worker/pyproject.toml
[project]
name = "buun-curator-worker"
version = "0.1.0"
description = "Temporal-based worker for Buun Curator feed reader"
requires-python = ">=3.12"
dependencies = [
    "temporalio>=1.21.1",
    "httpx>=0.28.1",
    "aiohttp>=3.13.2",
    "crawl4ai>=0.7.8",
    "feedparser>=6.0.12",
    "mcp>=1.25.0",
    "toon-format>=0.9.0b1",
    "aiobotocore>=3.0.0",
    "pillow>=11.3.0",
    "markdown>=3.10",
    "python-ulid>=3.1.0",
    "langchain>=1.2.0",
    "langchain-openai>=1.1.6",
    "githubkit>=0.14.1",
    "url-normalize>=2.2.1",
    "meilisearch>=0.38.0",
    "protego>=0.5.0",
    "graphiti-core[falkordb]>=0.25.0",
    "fastembed>=0.7.4",
    "lightrag-hku>=1.4.9",
    "ragas>=0.4.2",
    "langfuse>=3.11.2",
    "langchain-huggingface>=1.2.0",
    "sentence-transformers>=5.2.0",
    "rapidfuzz>=3.14.3",
    "opentelemetry-api>=1.39.1",
    "opentelemetry-sdk>=1.39.1",
    "opentelemetry-exporter-otlp-proto-grpc>=1.39.1",
    "structlog>=25.5.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.14.10",
    "pyright>=1.1.408",
    "types-aiobotocore[s3]>=3.0.0",
    "watchfiles>=1.1.1",
    "click>=8.3.1",
    "postgres-mcp>=0.3.0",
]

[tool.uv.sources]
toon-format = { git = "https://github.com/toon-format/toon-python.git" }

[project.scripts]
worker = "buun_curator.worker:main"
trigger = "scripts.trigger_workflow:cli"
schedule = "scripts.schedule:cli"
fetch = "scripts.fetch_content:cli"
generate-filtering-dataset = "scripts.generate_filtering_dataset:main"
run-filtering-on-dataset = "scripts.run_filtering_on_dataset:main"
generate-singlehop-eval-dataset = "scripts.generate_singlehop_eval_dataset:cli"
generate-multihop-eval-dataset = "scripts.generate_multihop_eval_dataset:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["buun_curator", "scripts"]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # Pyflakes (undefined names, unused imports, etc.)
    "I",    # isort (import sorting)
    "UP",   # pyupgrade (Python version upgrade suggestions)
    "B",    # flake8-bugbear (common bugs and design issues)
    "C4",   # flake8-comprehensions (unnecessary list/dict/set comprehensions)
    "SIM",  # flake8-simplify (simplifiable code patterns)
    "ARG",  # flake8-unused-arguments
]
ignore = []
extend-select = ["E501"]
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.pytest.ini_options]
asyncio_mode = "auto"
filterwarnings = [
    "ignore::pydantic.warnings.PydanticDeprecatedSince20",
    "ignore::DeprecationWarning:bs4.element",
]
markers = [
    "integration: marks tests as integration tests (deselect with '-m \"not integration\"')",
]

[tool.pyright]
pythonVersion = "3.12"
typeCheckingMode = "standard"
reportUnusedVariable = "warning"
venvPath = "."
venv = ".venv"
